requirements.txt
numba==0.53.0
numpy==1.23.5
nuscenes-devkit==1.1.11
mmcv-full==1.6.0
-f https://download.openmmlab.com/mmcv/dist/cu111/torch1.9.0/index.html
mmdet==2.28.2
mmdet3d==v1.0.0rc6
mmsegmentation==0.30.0
pexpect
yapf==0.40.1
setuptools==59.5.0